Switch detection
Added two switch blueprints, the main one and the secondary one. Created a blueprint interface and implemented it. Created a linetrace that tests on tick if a switch is in front of the player with DoesImplementInterface, and if so shows a UI telling the player to use the F key to interact (no interaction yet!)